Moved the summary of information to a seperate component
