+++ src/components/ReservationPageContainer.js
@@ -5,6 +5,7 @@ import { Link, useParams } from "react-router-dom";
 import ReservationPageCatering from "./ReservationPageCatering";
 import ReservationPageDayPart from "./ReservationPageDayPart";
 import ReservationPageSeatPlans from "./ReservationPageSeatPlans";
+import ReservationPageSummary from "./ReservationPageSummary";
 
 import LoadingSpinner from "./assets/LoadingSpinner";
 import LargeButton from "./assets/LargeButton";
@@ -86,10 +87,6 @@ const ReservationPageContainer = () => {
     padding: "15px",
   };
 
-  const divSelectedOptionsStyle = {
-    height: "270px",
-  };
-
   const leftStyle = {
     marginRight: "10px",
   };
@@ -100,38 +97,6 @@ const ReservationPageContainer = () => {
     flexDirection: "column",
     justifyContent: "space-between",
     boxShadow: "0 0 0 1px rgba(0,0,0,.15), 0 2px 3px rgba(0,0,0,.2)",
-  };
-
-  const inputFormStyle = {
-    padding: "0 10px 10px 10px",
-    textAlign: "center",
-  };
-
-  const ulStyle = {
-    listStyle: "none",
-    padding: "5px",
-    width: "60%",
-    margin: "0 auto",
-    textAlign: "center",
-  };
-
-  const liStyle = {
-    color: "#ed008c",
-    backgroundColor: "#fff",
-    border: "2px solid #ed008c",
-    borderRadius: "3px",
-    margin: "5px 0 5px 0",
-    padding: "5px",
-  };
-
-  const h4Style = {
-    margin: "5px 0 5px 0",
-  };
-
-  const pStyle = {
-    textAlign: "center",
-    fontSize: "14px",
-    color: "#636363",
   };
 
   const imageStyle = {
@@ -166,39 +131,16 @@ const ReservationPageContainer = () => {
       </div>
       <div style={rightStyle}>
         <DatePicker pickedDate={pickedDate} setPickedDate={setPickedDate} />
-        <div>
-          <div style={divSelectedOptionsStyle}>
-            <p style={pStyle}>Geselecteerde opties</p>
-            <ul style={ulStyle}>
-              {pickedDayPart && <li>Dagdeel</li>}
-              {pickedDayPart && <li style={liStyle}>{pickedDayPart}</li>}
-              {pickedSeatPlan && <li>Opstelling</li>}
-              {pickedSeatPlan && <li style={liStyle}>{pickedSeatPlan}</li>}
-              {pickedMeal && <li>Lunch</li>}
-              {pickedMeal && <li style={liStyle}>{pickedMeal}</li>}
-              {pickedExtraCatering && <li>Extra's</li>}
-              {pickedExtraCatering && <li style={liStyle}>{pickedExtraCatering}</li>}
-            </ul>
-          </div>
-          <p style={pStyle}>- - - - - - - - - - - - - - - - - - - - - - - - - - -</p>
-          <div style={inputFormStyle}>
-            <h4 style={h4Style}>Email:</h4>
-            <input type="email"></input>
-            <br></br>
-            <br></br>
-            <h4 style={h4Style}>Bedrag Schatting</h4>
-            <u>
-              <li>Zaal: €{pickedDayPart === "Hele dag" ? locationPrice * 2 : locationPrice}</li>
-              <li>Catering: €{mealPrice}</li>
-              <li>
-                Totaal: €
-                {(pickedDayPart === "Hele dag" ? locationPrice * 2 : locationPrice) + mealPrice}
-              </li>
-            </u>
 
-            <p style={pStyle}>Op basis van uw selectie wordt een offerte gemaakt</p>
-          </div>
-        </div>
+        <ReservationPageSummary
+          pickedDayPart={pickedDayPart}
+          pickedSeatPlan={pickedSeatPlan}
+          pickedMeal={pickedMeal}
+          pickedExtraCatering={pickedExtraCatering}
+          mealPrice={mealPrice}
+          locationPrice={locationPrice}
+        />
+
         <LargeButton text="Reserveren" />
       </div>
     </div>
